backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "commontalk-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "Flask",
  "gunicorn",
  "flask-cors",
  "psycopg2-binary",
  "requests",
  "openai",
  "groq",
  "scikit-learn",
  "dotenv",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["app*", "modules*"]
